+++ R/get_state_data.r
@@ -10,7 +10,7 @@
 #'
 #' @param years a vector of years to retrieve data for (i.e. \code{2014} for CDC
 #'        flu season 2014-2015). Default value is the current year and all
-#'        \code{years} values should be > \code{1997}
+#'        \code{years} values should be >= \code{2008}
 #' @return A \code{data.frame} of state-level data for the specified seasons
 #'         (also classed as \code{cdcstatedata})
 #' @export
@@ -26,34 +26,22 @@
 #' }
 get_state_data <- function(years=as.numeric(format(Sys.Date(), "%Y"))) {
 
-  if (any(years < 1997))
-    stop("Error: years should be > 1997")
+  if (any(years < 2008))
+    stop("Error: years should be >= 2008")
 
-  years <- years - 1960
+  years <- c((years - 1960), 1)
+  years <- paste0(years, collapse=",")
 
-  out_file <- tempfile(fileext=".zip")
-
-  params <- list(EndMMWRID=0,
-                 StartMMWRID=0,
-                 QueryType=1,
-                 DataMode="STATE",
-                 SeasonsList=paste0(years, collapse=","))
-
-  tmp <- httr::POST("https://gis.cdc.gov/grasp/fluview/FluViewPhase1CustomDownload.ashx",
-              body=params,
-              write_disk(out_file))
+  tmp <- httr::GET(sprintf("https://gis.cdc.gov/grasp/fluView1/Phase1DownloadDataP/%s", years))
 
   stop_for_status(tmp)
 
-  if (!(file.exists(out_file)))
-    stop("Error: cannot process downloaded data")
+  res <- httr::content(tmp, as="parsed")
 
-  out_dir <- tempdir()
-
-  files <- unzip(out_file, exdir=out_dir, overwrite=TRUE)
-
-    out <- read.csv(files, header=TRUE, stringsAsFactors=FALSE)
-  # out <- readr::read_csv(files, col_types="cccccic")
+  ctx <- v8()
+  ctx$eval(JS(sprintf("var dat=%s;", res)))
+  res <- ctx$get("dat", flatten=FALSE)
+  out <- suppressMessages(readr::type_convert(res$datadownload))
 
   class(out) <- c("cdcstatedata", class(out))