+++ NEWS.md
@@ -1,6 +1,7 @@
 # cscfluview 0.5.1
 
 * Replaced `http` URLs with `https` as `http` ones no longer work (fixes #6)
+* Fixed State data download (CDC changed the hidden API)
 
 # cdcfluview 0.5.0